Compatibility for upcoming PlaylistManager
The upcoming PlaylistManager (and BeatSaberPlaylistsLib) will be using a new way of loading levels into the game to make them more transparent to other mods (as playlist songs are their own type with their own unique data and PlaylistManager needs to access it). As a result, SongBrowser will need to get the songs directly as PlaylistSongs to maintain compatibility. DM me on Discord if you have any questions.